fix: NoticeContent 컴포넌트 분리 및 DropDown 컴포넌트 라이프사이클 이슈 - #128
